// src/chains/functions/[Harmony-32] Tranquil Finance - Lending Deposit/data/interfaces.ts
export interface Contracts {
  tqERC20DelegatorUSDCAddress: string;
  tqERC20DelegatorONEAddress: string;
  tqERC20DelegatorstONEAddress: string;
  tqERC20Delegator1WBTCAddress: string;
  tqERC20Delegator1BTCAddress: string;
  tqERC20Delegator1ETHAddress: string;
  tqERC20Delegator1USDTAddress: string;
  tqERC20Delegator1DAIAddress: string;
  "1USDCAddress": string;
  stONEAddress: string;
  "1WBTCAddress": string;
  "1BTCAddress": string;
  "1ETHAddress": string;
  "1USDTAddress": string;
  "1DAIAddress": string;
}

export interface OTFSettings {
  target: string;
  qtyType: string;
  qty: number;
}